+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.exception.UserNotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.service.InMemoryUserService;
 import ru.yandex.practicum.filmorate.service.UserService;
 
 import java.util.*;
@@ -15,9 +16,12 @@ import java.util.*;
 @Slf4j
 @Component
 public class UserController {
+    UserService userService;
 
     @Autowired
-    UserService userService;
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
 
     @PostMapping("/users")
     public User addUser(@RequestBody User user) throws ValidationException {
@@ -38,34 +42,34 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
-    public User getUserById(@PathVariable int id) {
+    public User getUserById(@PathVariable long id) {
         log.debug("Запрос на получение данных конкретного пользователя {}", id);
         return userService.getUserById(id);
     }
 
     @PutMapping("/users/{id}/friends/{friendId}")
-    public void addFriend(@PathVariable int id,
-                          @PathVariable int friendId) {
+    public void addFriend(@PathVariable long id,
+                          @PathVariable long friendId) {
         log.debug("Запрос на добавление в друзья пользователей {} и {}", id, friendId);
         userService.addFriend(id, friendId);
     }
 
     @DeleteMapping("/users/{id}/friends/{friendId}")
-    public void deleteFriend(@PathVariable int id,
-                          @PathVariable int friendId) {
+    public void deleteFriend(@PathVariable long id,
+                          @PathVariable long friendId) {
         log.debug("Запрос на удаление из друзей пользователей {} и {}", id, friendId);
         userService.deleteFriend(id, friendId);
     }
 
     @GetMapping ("/users/{id}/friends")
-    public List<User> getFriendsOfUserId(@PathVariable int id) {
+    public List<User> getFriendsOfUserId(@PathVariable long id) {
         log.debug("Запрос на список друзей пользователя {}", id);
         return userService.getFriendsOfUserId(id);
     }
 
     @GetMapping ("/users/{id}/friends/common/{otherId}")
-    public Set<User> getCommonFriends (@PathVariable int id,
-                                          @PathVariable int otherId) {
+    public Set<User> getCommonFriends (@PathVariable long id,
+                                          @PathVariable long otherId) {
         log.debug("Запрос на список общих друзей пользователей {} и {}", id, otherId);
         return userService.getCommonFriends(id, otherId);
     }
