+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -11,7 +11,7 @@ public interface FilmStorage {
 
     List<Film> getFilms();
 
-    Film getFilmById(int id);
+    Film getFilmById(long id);
 
     boolean checkValidity(Film film);
 }
